requirements.txt
apache-airflow==2.9.1
requests
sqlite3 # comes with Python, just for clarity